requirements.txt
python-dotenv
fastapi
uvicorn
sqlalchemy
asyncpg
fastapi-cache2[redis]
gunicorn
fastapi-users[sqlalchemy]